Somewhat better fix for readonly directories.
The previous approach to readonly files still maintained:

os.MkdirAll(dst, d.info.Mode())

The problem with this is that if dst is a/b/c and the src directory
is readonly... then MkdirAll creates:

a/ - readonly
then tries to create a/b

but as a/ is readonly... fails with a permissions error.

This change switches to:

os.MkdirAll(dst, d.info.Mode()|os.ModePerm)

which basically creates the directory (and any of its ancestors)
with the maximum permissions allowable by the users umask.

Then after all the recursive creating of things we use:

os.MkdirAll(dst, d.info.Mode())

makes sure that the dst directory itself (and not its ancestors created)
incidentally) to the permissions of the original src directory.

Signed-off-by: Ed Warnicke <hagbard@gmail.com>