Add proto file to project
Proto file is a copy of the one used in formation firmware project.
Proto file will now have an additional field "identifier" under the
FixtureToTower message. This message is a string and will be used to
indicate which Fixture is sending out the message. Along with the proto
file the compiled proto.pb.go file is also commited.