Ignore mismatch from fixture barcode in proto
We keep the state internally, so don't fail out if the FXR barcode
doesn't match what we have configured. This is currently an
inconsistency in the FXR. We can add this back when we trust the FXR to
broadcast what we expect consistently.

The current case where we see this issue is with an FXR that is
reporting ACTIVE with no tray in it. This causes the TC SM to jump to
IN_PROGRESS, but there's no tray so the FXR is not reporting a barcode.

Removing this allows the TC to be more robust to FXR issues.