Unreserve FXR after a timeout when tray does not arrive
For now set to a timeout of 10 minutes. At C-Tower this is a very long
time. This may have to change (or be removed entirely) for mass
production but for now it will allow us to cycle through bad recipes or
missteps from other parts of the system while validating.

TF-1222