Read configuration upon changes to the file
This required some weird checks in the configuration monitor due to some
apparent weirdness in how the docker base OS is reporting file changes
on shared volumes. Basically the last step in a write is a remove
(following a rename and chmod). Because of this we can't just look for
WRITE events, but also for RENAME events. Because the inode changes we
then also need to remove the old watcher and add a new one (albeit with
the same name).

This update allows us to change configuration on the fly, but protects
currently-running state machines from updating configuration except
while they are idling with no tray. This seems the safer way to do it.

A simple global variable was used to hold the master configuration copy.
It is copied into each SM prior to use.

This also prevents the configuration from advancing out of IDLE if it is
not allowed.

TF-1152