Remove reservation timeout as it is no longer needed
Now that manual reservation reset has been merged we no longer need, or
want, this feature. It turns out that when the crane faults and comes
back after the timeout we can end up sending to the same fixture twice
when the crane comes back online.

Remove the timeout. If reservations need to be reset we can remove them
manually via the endpoint.

TF-1222