Add generic form_request and equipment_request endpoints
These two endpoints allow a user to POST requests to the TC to pass
along a FormRequest or EquipmentRequest. The form_request endpoint
replaces the fault reset endpoint and allows the user to do the same
thing. The equipment_request endpoint allows the user to send a self
test approval (or any other EquipmentRequeset) to an FXR.

TF-997