Add /avail query to look at all fixtures
Instead of just reporting the allowed fixtures report all fixtures by
default. Add a query parameter ?allowed that, if set to true, only
reports the allowed fixtures.

Because normal operation is changing to report all fixtures reduce the
timeout from 3 to 2 seconds so the response time isn't so long.

Remove a log field (response body) because it was just printing the
bytes.

TF-1150