Unify fixture reads to background read/update
Instead of explicitly reading off the channel every time we need to get
a new value from the fixture, run that process in the background so the
state is constantly being updated. Then access the fixture status just
by accessing that data. The getters invalidate the data if it has gone
stale.