Add update endpoint to schedule update on service
Add endpoint to schedule an update on the towercontroller. POSTing to
this endpoint reserves all fixtures until they are all idle again then
exits with an error code so towercontroller restarts. This triggers
systemd to pull the latest docker image and restart it automatically (if
towercontroller is installed correctly). The fixtures are reserved by
ORing their reservation status with an _updateScheduled flag.

A GET to this endpoint returns true/false indicating whether an update
has been scheduled.

An additional endpoint allows a user to cancel a scheduled update.

TF-1276, TF-1277