Add lock to load request handler
This lock prevents a race on the check for fixture status. If a TC gets
a double load request previously both threads handling the request would
race to validate fixture status before updating fixture status to a
blocking status.

Add a "debounce" after writing to the channel in order to allow the
fixture to update its state.

TF-1636