Do not request unload when fire alarm triggered
Although, due to sequence of events, the unload is triggered after the
fire response is commanded, adding an unload job to the job queue could
cause difficulty recovering after a fire event. When a fire alarm is
triggered log the fault and return to idle. Let the watchtower service
and crane handle the unload.