Turn off log duplication by docker
The default log driver of the docker daemon is JSON. These logs get
written to STDOUT which gets picked up by the docker JSON logs and by
syslog. All TC services write their own logs to /var/log so this
duplication is not necessary and causes disk bloat if not careful.

Set the log-driver to none so the only copy of the logs is under
/var/log.