Add fault recovery to tray transactions
Query C/D Controller for fault record with ops snapshot before sending
recipe information to the FXR. This ops snapshot will be sent to the FXR
so the FXR can pick up where it left off.

If a fixture faults send a fault record to C/D Controller so it can be
queried by the next fixture that tries to run the tray.

If a fixture passes the recipe POST a clear request to the C/D
controller to remove the fault record.

TF-1737