Egress trays to inspection station when fault limit reached
When a tray has faulted too many fixtures place a hold on it so it gets
routed to cm_inspect. This is to protect against a tray faulting too
many fixtures without being removed from the system.

Also do not post a fault record for commission-self-test trays, as these
are likely to fault a lot of fixtures.