Add tray orientation to TowerToFixture message
The FXR will then broadcast the orientation back in the Operational
message. On the event of a fault the orientation will be recorded in the
operational snapshot. When a fixture resumes with that tray it will be
able to pull previous orientation from the snapshot and new orientation
from the new TowerToFixture message, and apply the saved cell statuses
correctly.

TF-1756