Add to data expiry timeout
Operational messages are increasing from 1Hz to .2Hz when fixture is
Idle. Correspondingly update the data expiry to 40% above highest data
transmission delay.

This is the simple approach. An alternative approach would be to
increase data expiry only when the fixture is in IDLE. This method comes
with risks and future required changes if data rate is changed for other
states. Opted for simple approach for now.

TF-1777