Add protections to duplicate messages from CDC
Sometimes WMS sends duplicate messages to CND despite acks working
properly. Because CND and CDC both attempt to be as stateless as
possible these are passed directly down the stack without duplication
validation. TC is a stateful design and therefore it is necessary to do
this validation.

Effectively: only accept one prepared_for_delivery and one load request
at a time. Discard duplicates or any messages that come through when the
fixture is unavailable.

TF-1636, TF-1759