Add automatic maintenance request for fixture timeouts
There are two timeouts that can occur when sending recipes to a fixture.

The first part of the handshake is a wait for the fixture to report
READY, indicating a tray is in the fixture, before sending the recipe.
If the fixture fails to report READY after a minute tower controller
assumes the tray present sensor isn't working and requests an unload.

The second part of the handshake is the actual attempt to send the
recipe. Tower controller sends the recipe along with a transaction ID,
then checks if the transaction ID broadcast by the fixture matches what
was sent. If it does not match what was sent it continues sending on a
cadence of once per two seconds (one second to write, one second to
read). If the transaction ID doesn't update within one minute tower
controller assumes the fixture is not communicating properly and
requests an unload.

In the first case automatically request maintenance with a maintenance
reason of tray present sensor issue. In the second case automatically
request maintenance with a maintenance reason of fixture not responding.

TF-2482