Reduce fault retries to one
Production has requested the number of retries when a tray faults a
fixture be reduced to only one retry, making the max fixtures a tray can
fault be two fixtures.

This request is due to recent poorly-constructed recipes causing fixture
faults. Since the recipes were applied to a large number of trays and
each tray faulted four fixtures before the hold was placed this caused
high fixture fallout on the line.

The original reason for three retries was due to occasional misalignment
when placing a tray into a fixture. The system would then naturally try
to find a new place to put the tray and (hopefully) place with correct
alignment this time. Misalignment is no longer the large problem it used
to be, so egressing a tray after two faults will not overload the
production team.

TF-3281